Move command.go to subpkg to fix build breakage on darwin
